Split Gateway and VirtualService resources to separate files.
