+++ nca/Parsers/IstioSidecarYamlParser.py
@@ -6,8 +6,7 @@ import re
 from nca.CoreDS.Peer import PeerSet
 from nca.Resources.NetworkPolicy import NetworkPolicy
 from nca.Resources.IstioSidecar import IstioSidecar, IstioSidecarRule
-from nca.Resources.IstioTrafficResources import istio_root_namespace
-from .IstioGenericYamlParser import IstioGenericYamlParser
+from .IstioGenericYamlParser import IstioGenericYamlParser, istio_root_namespace
 
 
 class IstioSidecarYamlParser(IstioGenericYamlParser):
