+++ nca/Resources/VirtualService.py
@@ -9,69 +9,6 @@ from nca.CoreDS.PortSet import PortSet
 from nca.CoreDS.MinDFA import MinDFA
 
 
-# TODO: istio_root_namespace should be configurable from istio configuration, currently using default value for it
-# If namespace is set to istio root namespace, the policy object applies to all namespaces in a mesh
-istio_root_namespace = 'istio-config'
-
-
-class Gateway:
-    """
-    A class for keeping some elements of parsed Istio Gateway, needed for building IstioGatewayPolicy
-    """
-
-    @dataclass
-    class Server:
-        @dataclass
-        class GatewayPort:
-            number: int
-            protocol: str
-            name: str
-
-        port: GatewayPort  # the port field is not currently  used.
-        hosts_dfa: MinDFA or None = None
-        name: str = ''
-
-        def add_host(self, host_dfa):
-            """
-            Add host_dfa to the server
-            :param MinDFA host_dfa:
-            :return:
-            """
-            if self.hosts_dfa:
-                self.hosts_dfa |= host_dfa
-            else:
-                self.hosts_dfa = host_dfa
-
-    def __init__(self, name, namespace):
-        """
-        Create a Gateway
-        :param str name: the gateway name
-        :param K8sNamespace namespace: the gateway namespace
-        """
-        self.name = name
-        self.namespace = namespace
-        self.peers = PeerSet()
-        self.servers = []
-        self.all_hosts_dfa = None
-
-    def full_name(self):
-        """
-        :return str: the full gateway name in <namespace>/<name> format
-        """
-        return str(self.namespace) + '/' + self.name
-
-    def add_server(self, server):
-        """
-        Add a server and collect its host_dfa into one dfa
-        :param Gateway.Server server: a server to add
-        """
-        self.servers.append(server)
-        if self.all_hosts_dfa:
-            self.all_hosts_dfa |= server.hosts_dfa
-        else:
-            self.all_hosts_dfa = server.hosts_dfa
-
-
 class VirtualService:
     """
     A class for keeping some elements of parsed Istio VirtualService, needed for building IstioGatewayPolicy
